PIM-5094: Move common logic for import/export in a dedicated class
